# Add effective index to subscriptions