Optimize Pubsubhubbub::DistributionWorker

IMO this optimization is not important for most instance owner because Subscription.count is maybe small. However all we have to do is only reverse indexes.

```
SELECT id, callback_url FROM "subscriptions" WHERE "subscriptions"."account_id" = X AND "subscriptions"."confirmed" = 't' AND ("subscriptions"."expires_at" > '2017-05-07 14:16:36.159176')
```

**before**

```
Seq Scan on subscriptions  (cost=0.00..5102.81 rows=342 width=48) (actual time=0.042..21.021 rows=264 loops=1)
  Filter: (confirmed AND (expires_at > '2017-05-07 14:16:36.159176'::timestamp without time zone) AND (account_id = X))
  Rows Removed by Filter: 138723
Planning time: 0.113 ms
Execution time: 21.039 ms
```

**after**

```
Bitmap Heap Scan on subscriptions  (cost=24.92..1497.97 rows=394 width=48) (actual time=0.097..0.461 rows=264 loops=1)
  Recheck Cond: (account_id = X)
  Filter: (confirmed AND (expires_at > '2017-05-07 14:16:36.159176'::timestamp without time zone))
  Rows Removed by Filter: 229
  Heap Blocks: exact=382
  ->  Bitmap Index Scan on index_subscriptions_on_account_id_and_callback_url  (cost=0.00..24.82 rows=587 width=0) (actual time=0.060..0.060 rows=493 loops=1)
        Index Cond: (account_id = X)
Planning time: 0.048 ms
Execution time: 0.488 ms
```